+++ CHANGES.md
@@ -38,6 +38,8 @@ To be released.
     [[#51]]
  -  Added `Bencodex.Types.Dictionary[Text]` indexer. [[#52]]
  -  Added `Bencodex.Types.Dictionary[Binary]` indexer. [[#52]]
+ -  Added `Bencodex.Types.Dictionary.ContainsKey(Text)` method.  [[#52]]
+ -  Added `Bencodex.Types.Dictionary.ContainsKey(Binary)` method.  [[#52]]
  -  Added `Bencodex.Misc.KeyComparer` class.  [[#51]]
  -  `List.Empty` static property became a static readonly field.  [[#50]]
  -  `Dictionary.Empty` static property became a static readonly field.  [[#50]]